refactor(server): moving common base models to services layer
Also update module imports
